Reworks the maintenance link. Now it opens a dialog to allow individual deletions.
git-svn-id: svn://cherokee-project.com/cherokee/trunk@5698 5dc97367-97f1-0310-9951-d761b3857238

+++ admin/market/Maintenance.py
@@ -29,6 +29,7 @@ import Page
 import stat
 import time
 import popen
+import string
 import OWS_Login
 
 from util import *
@@ -36,9 +37,15 @@ from consts import *
 from ows_consts import *
 from configured import *
 
+NOTE_DBUSER = N_("Specify the name of a privileged DB user. It will be used to automatically create the required database.")
+NOTE_DBPASS = N_("Specify the password for this user account.")
+DB_DEL_H2   = N_("Database removal")
+DB_DEL_P1   = N_("Please, provide an administrative user/password pair to connect to your database.")
+
 URL_MAINTENANCE_LIST       = "/market/maintenance/list"
 URL_MAINTENANCE_LIST_APPLY = "/market/maintenance/list/apply"
-
+URL_MAINTENANCE_DB         = "/market/maintenance/db"
+URL_MAINTENANCE_DB_APPLY   = "/market/maintenance/db/apply"
 
 #
 # Cache
@@ -75,7 +82,7 @@ def check_unfinished_installations():
     for d in os.listdir (CHEROKEE_OWS_ROOT):
         fd = os.path.join (CHEROKEE_OWS_ROOT, d)
 
-        if not all([c in "0123456789." for c in d]):
+        if not all([c in string.digits for c in d]):
             continue
 
         if not os.path.isdir(fd):
@@ -106,7 +113,7 @@ def check_orphan_installations():
     for d in os.listdir (CHEROKEE_OWS_ROOT):
         fd = os.path.join (CHEROKEE_OWS_ROOT, d)
 
-        if not fd in cfg_txt:
+        if not re.match (r'document_root.*'+fd, cfg_txt):
             orphans.append(d)
 
     orphan_installs_cache = orphans
@@ -123,13 +130,61 @@ def does_it_need_maintenance():
     return False
 
 
-def remove_app_databse (app):
-    None
+def app_database_exists (app):
+    fp = os.path.join (CHEROKEE_OWS_ROOT, app, "install.log")
+    cont = open(fp, 'r').read()
+
+    # MySQL
+    if 'DB: Created MySQL database' in cont:
+        return "MySQL"
+
+
+def app_database_remove (app, user, passw, db_type):
+    if db_type == 'MySQL':
+        cmd = "mysql -u'%(user)s' -p'%(passw)s' -e 'DROP DATABASE IF EXISTS market_%(app)s'" %(locals())
+
+        ret = popen.popen_sync (cmd)
+        if ret['retcode'] != 0:
+            return ret['stderr']
 
 
 #
 # GUI
 #
+
+class Maintenance_Box (CTK.Box):
+    def __init__ (self, refresh):
+        CTK.Box.__init__ (self)
+
+        unfinished = check_unfinished_installations()
+        orphan     = check_orphan_installations()
+
+        if not len(unfinished) and not len(orphan):
+            return
+
+        dialog = MaintenanceDialog()
+        self += CTK.RawHTML ('<h3>%s</h3>' %(_('Maintanance')))
+        self += dialog
+
+        if len(unfinished):
+            link = CTK.Link (None, CTK.RawHTML(_('Clean up')))
+            link.bind ('click', dialog.JS_to_show())
+
+            box = CTK.Box()
+            box += CTK.RawHTML ('%d %s' %(len(unfinished), _("partial installations: ")))
+            box += link
+
+            self += box
+
+        if len(orphan):
+            link = CTK.Link (None, CTK.RawHTML(_('Clean up')))
+            link.bind ('click', dialog.JS_to_show())
+
+            box = CTK.Box()
+            box += CTK.RawHTML ('%d %s' %(len(orphan), _("orphan installations: ")))
+            box += link
+            self += box
+
 
 class MaintenanceDialog (CTK.Dialog):
     def __init__ (self):
@@ -150,7 +205,7 @@ class AppList (CTK.Table):
     def __init__ (self, apps):
         CTK.Table.__init__ (self, {'id': 'maintenance-removal-list'})
 
-        self += [CTK.RawHTML(x) for x in (_('Remove'), _('Application'), _('Status'), _('Date'))]
+        self += [CTK.RawHTML(x) for x in (_('Remove'), _('Application'), _('Status'), _('Database'), _('Date'))]
         self.set_header()
 
         for app in apps:
@@ -158,6 +213,7 @@ class AppList (CTK.Table):
             self += [check,
                      CTK.RawHTML (apps[app]['name']),
                      CTK.RawHTML (apps[app]['type']),
+                     CTK.RawHTML (apps[app].get('db', '')),
                      CTK.RawHTML (apps[app]['date'])]
 
 
@@ -167,22 +223,37 @@ class ListApps:
         remove_apps = {}
 
         for app in check_orphan_installations():
+            db = app_database_exists (app)
+
             remove_apps[app] = {}
             remove_apps[app]['type'] = _('Orphan application')
             remove_apps[app]['name'] = self._figure_app_name (app)
             remove_apps[app]['date'] = self._figure_app_date (app)
+            if db:
+                remove_apps[app]['db'] = db
 
         for app in check_unfinished_installations():
             if not app in remove_apps:
+                db = app_database_exists (app)
+
                 remove_apps[app] = {}
                 remove_apps[app]['type'] = _('Unfinished installation')
                 remove_apps[app]['name'] = self._figure_app_name (app)
                 remove_apps[app]['date'] = self._figure_app_date (app)
+                if db:
+                    remove_apps[app]['db'] = db
+
+        # Store in CTK.cfg
+        del (CTK.cfg ['admin!market!maintenance!remove'])
+        for app in remove_apps:
+            CTK.cfg ['admin!market!maintenance!remove!%s!del' %(app)] = 0
+            CTK.cfg ['admin!market!maintenance!remove!%s!db'  %(app)] = remove_apps[app].get('db')
+            CTK.cfg ['admin!market!maintenance!remove!%s!name'%(app)] = remove_apps[app]['name']
 
         # Dialog buttons
         buttons  = CTK.DruidButtonsPanel()
         buttons += CTK.DruidButton_Close (_('Cancel'))
-        buttons += CTK.DruidButton_Submit (_('Remove'))
+        buttons += CTK.DruidButton_Goto  (_('Next'), URL_MAINTENANCE_DB, True)
 
         # Content
         cont  = CTK.Container()
@@ -227,14 +298,17 @@ def ListApps_Apply():
             if CTK.post.get_val(k) == '1':
                 apps_to_remove.append (app)
 
+    if not apps_to_remove:
+        return {'ret': 'fail'}
+
+    # Store databases to remove
+    for n in range (len(apps_to_remove)):
+        CTK.cfg ['admin!market!maintenance!remove!%s!del' %(app)] = '1'
+
     # Perform the app removal
     for app in apps_to_remove:
-        # App Document root
         fp = os.path.join (CHEROKEE_OWS_ROOT, app)
         popen.popen_sync ("rm -rf '%s'" %(fp))
-
-        # Database removal
-        remove_app_databse (app)
 
     # The cache is no longer valid
     Invalidate_Cache()
@@ -242,40 +316,60 @@ def ListApps_Apply():
     return CTK.cfg_reply_ajax_ok()
 
 
-class Maintenance_Box (CTK.Box):
-    def __init__ (self, refresh):
-        CTK.Box.__init__ (self)
+class DatabaseRemoval:
+    def __call__ (self):
+        # Check the application
+        db_found = False
 
-        unfinished = check_unfinished_installations()
-        orphan     = check_orphan_installations()
+        for app in CTK.cfg.keys ('admin!market!maintenance!remove'):
+            db_type = CTK.cfg.get_val ('admin!market!maintenance!remove!%s!db'%(app))
+            if db_type:
+                db_found = True
+            else:
+                del (CTK.cfg ['admin!market!maintenance!remove!%s'%(app)])
 
-        if not len(unfinished) and not len(orphan):
-            return
-
-        dialog = MaintenanceDialog()
-        self += CTK.RawHTML ('<h3>%s</h3>' %(_('Maintanance')))
-        self += dialog
-
-        if len(unfinished):
-            link = CTK.Link (None, CTK.RawHTML(_('Clean up')))
-            link.bind ('click', dialog.JS_to_show())
-
+        # No DBs, we are done here
+        if not db_found:
             box = CTK.Box()
-            box += CTK.RawHTML ('%d %s' %(len(unfinished), _("partial installations: ")))
-            box += link
+            box += CTK.RawHTML (js = CTK.DruidContent__JS_to_close (box.id))
+            return box.Render().toStr()
 
-            self += box
+        # Ask for the user and password
+        table = CTK.PropsTable()
+        table.Add (_('DB user'),     CTK.TextField        ({'name':'db_user', 'class':'noauto', 'value':'root'}), _(NOTE_DBUSER))
+        table.Add (_('DB password'), CTK.TextFieldPassword({'name':'db_pass', 'class':'noauto'}), _(NOTE_DBPASS))
 
-        if len(orphan):
-            link = CTK.Link (None, CTK.RawHTML(_('Clean up')))
-            link.bind ('click', dialog.JS_to_show())
+        submit  = CTK.Submitter (URL_MAINTENANCE_DB_APPLY)
+        submit.bind ('submit_success', CTK.DruidContent__JS_to_close (table.id))
+        submit += CTK.Hidden ('db_type', db_type)
+        submit += table
 
-            box = CTK.Box()
-            box += CTK.RawHTML ('%d %s' %(len(orphan), _("orphan installations: ")))
-            box += link
+        cont  = CTK.Container()
+        cont += CTK.RawHTML ('<h2>%s</h2>' %(_(DB_DEL_H2)))
+        cont += CTK.RawHTML ('<p>%s</p>'   %(_(DB_DEL_P1)))
+        cont += submit
 
-            self += box
+        return cont.Render().toStr()
+
+
+def DatabaseRemoval_Apply():
+    db_user = CTK.post.get_val('db_user')
+    db_pass = CTK.post.get_val('db_pass')
+    db_type = CTK.post.get_val('db_type')
+
+    for k in CTK.cfg.keys ('admin!market!maintenance!remove'):
+        app = CTK.cfg.get_val ('admin!market!maintenance!remove!%s'%(k))
+
+        error = app_database_remove (app, db_user, db_pass, db_type)
+        if error:
+            return {'ret': 'fail', 'errors': {'db_pass': error}}
+
+    del (CTK.cfg['admin!market!maintenance!remove'])
+    return CTK.cfg_reply_ajax_ok()
+
 
 
 CTK.publish ('^%s$'%(URL_MAINTENANCE_LIST),       ListApps)
 CTK.publish ('^%s$'%(URL_MAINTENANCE_LIST_APPLY), ListApps_Apply, method="POST")
+CTK.publish ('^%s$'%(URL_MAINTENANCE_DB),         DatabaseRemoval)
+CTK.publish ('^%s$'%(URL_MAINTENANCE_DB_APPLY),   DatabaseRemoval_Apply, method="POST")
